+++ service-implementation/src/main/java/com/liaison/mailbox/service/dto/configuration/request/RemoteProcessorPropertiesDTO.java
@@ -107,6 +107,7 @@ public class RemoteProcessorPropertiesDTO {
 		this.connectionTimeout = connectionTimeout;
 	}
 
+	@PatternValidation(errorMessage = "Invalid URL.", type = MailBoxConstants.PROPERTY_URL)
 	public String getUrl() {
 		return url;
 	}